+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+# v5.13.1
+## 🛠 Fixes
+- Fix in the toolbar measuring method for `AndesCoachmark` | Author: [@juanunzu](https://github.com/juanunzu)
+
 # v5.13.0
 ## 🚀 Feature
 - Now the `AndesTextView` can be used as label for both `AndesTextfield` and `AndesAmountFieldSimple` components | Author: [@juanunzu](https://github.com/juanunzu)
+- Added custom content in `AndesList`
 
 # v5.12.0
 ## 🚀 Feature